Attempt a start at a util for test metadata introspection
