+++ Cargo.toml
@@ -12,14 +12,15 @@ ordered-float = "5.0.0"
 random_color = { version = "1.0.0", optional = true }
 rerun = { version = "0.22.1", optional = true }
 serde = { version = "1.0.218", features = ["derive"] }
-serde_json = "1.0.139"
+serde_json = "1.0.140"
+struct-field-names-as-array = "0.3.0"
 walkdir = "2.5.0"
 
 [dev-dependencies]
 assert_approx_eq = "1.1.0"
 criterion = { version = "0.5", features = ["html_reports"] }
 paste = "*"
-rstest = "0.24.0"
+rstest = "0.25.0"
 rstest_reuse = "*"
 tempfile = "3.17.1"
 
@@ -34,6 +35,10 @@ path = "src/bin/generate_rotated_ipa_polygons.rs"
 name = "visualizer"
 path = "src/bin/run_visualizer.rs"
 required-features = ["visualizer"]
+
+[[bin]]
+name = "sync-test-case-attrs"
+path = "src/bin/sync_test_case_attributes.rs"
 
 [[bench]]
 name = "convex_hull"
